New - sigul::server::gpg_kludge parameter/feature
At present this kludge is needed in order to run Sigul on CentOS 7 using
packages from Fedora Infrastructure's builder-rpms repository.  It
effectively makes gpg1 the default version of gpg, which could easily
render other aspects of the OS broken.  However, since Sigul Servers are
to be very isolated in access and functionality this shouldn't present
much of a problem in reality.