Drop - gpg kludge support for older platforms
Sigul Server once required gpg1 and was incompatible with gpg2. Some
platforms had gpg2 as the default when summoned merely as "gpg". To make
those platforms work, this module provided this kludge support.

Version 1.0-alpha.1 gained gpg2 support and this module no longer aims
to support platforms with an older version, so this support is being
dropped.

Signed-off-by: John Florian <jflorian@doubledog.org>